# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BebeMart API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  both: "both",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  state?: string | null;
  city?: string | null;
  rating?: number | null;
  reviewCount: number;
  createdAt: string;
}

export interface UserPublic {
  id: number;
  name: string;
  state?: string | null;
  city?: string | null;
  rating?: number | null;
  reviewCount: number;
  createdAt: string;
}

export type ListingStatus = (typeof ListingStatus)[keyof typeof ListingStatus];

export const ListingStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export interface Listing {
  id: number;
  title: string;
  description: string;
  price: number;
  category: string;
  ageGroup: string;
  condition: string;
  images: string[];
  videoUrl?: string | null;
  sellerId: number;
  sellerName: string;
  sellerRating?: number | null;
  state: string;
  city: string;
  status: ListingStatus;
  giftItForward: boolean;
  createdAt: string;
}

export type ListingDetail = Listing & {
  seller: UserPublic;
};

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateListingInputCategory =
  (typeof CreateListingInputCategory)[keyof typeof CreateListingInputCategory];

export const CreateListingInputCategory = {
  Clothes: "Clothes",
  Shoes: "Shoes",
  Toys: "Toys",
  Gear: "Gear",
  Books: "Books",
} as const;

export type CreateListingInputAgeGroup =
  (typeof CreateListingInputAgeGroup)[keyof typeof CreateListingInputAgeGroup];

export const CreateListingInputAgeGroup = {
  "0-6m": "0-6m",
  "6-12m": "6-12m",
  "1-3y": "1-3y",
  "4-6y": "4-6y",
  "7y+": "7y+",
} as const;

export type CreateListingInputCondition =
  (typeof CreateListingInputCondition)[keyof typeof CreateListingInputCondition];

export const CreateListingInputCondition = {
  Like_New: "Like New",
  Good: "Good",
  Fair: "Fair",
} as const;

export interface CreateListingInput {
  title: string;
  description: string;
  /**
   * @minimum 0
   * @maximum 5000
   */
  price: number;
  category: CreateListingInputCategory;
  ageGroup: CreateListingInputAgeGroup;
  condition: CreateListingInputCondition;
  /** @minItems 1 */
  images: string[];
  videoUrl?: string | null;
  state: string;
  city: string;
  giftItForward: boolean;
}

export type UpdateListingInputStatus =
  (typeof UpdateListingInputStatus)[keyof typeof UpdateListingInputStatus];

export const UpdateListingInputStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export interface UpdateListingInput {
  title?: string;
  description?: string;
  /**
   * @minimum 0
   * @maximum 5000
   */
  price?: number;
  category?: string;
  ageGroup?: string;
  condition?: string;
  images?: string[];
  videoUrl?: string | null;
  state?: string;
  city?: string;
  status?: UpdateListingInputStatus;
}

export type TransactionEscrowStatus =
  (typeof TransactionEscrowStatus)[keyof typeof TransactionEscrowStatus];

export const TransactionEscrowStatus = {
  pending: "pending",
  paid_escrow: "paid_escrow",
  shipped: "shipped",
  received: "received",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Transaction {
  id: number;
  listingId: number;
  listingTitle: string;
  listingImages: string[];
  buyerId: number;
  buyerName: string;
  sellerId: number;
  sellerName: string;
  amount: number;
  platformFee: number;
  escrowStatus: TransactionEscrowStatus;
  deliveryAddress?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTransactionInput {
  listingId: number;
  deliveryAddress: string;
}

export type UpdateTransactionStatusInputStatus =
  (typeof UpdateTransactionStatusInputStatus)[keyof typeof UpdateTransactionStatusInputStatus];

export const UpdateTransactionStatusInputStatus = {
  shipped: "shipped",
  received: "received",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateTransactionStatusInput {
  status: UpdateTransactionStatusInputStatus;
}

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  content: string;
  listingId?: number | null;
  read: boolean;
  createdAt: string;
}

export interface SendMessageInput {
  content: string;
  listingId?: number | null;
}

export interface Conversation {
  otherUser: UserPublic;
  lastMessage: Message;
  unreadCount: number;
}

export interface Review {
  id: number;
  buyerId: number;
  buyerName: string;
  sellerId: number;
  transactionId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  createdAt: string;
}

export interface CreateReviewInput {
  sellerId: number;
  transactionId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
}

export type GuestLoginInputRole =
  (typeof GuestLoginInputRole)[keyof typeof GuestLoginInputRole];

export const GuestLoginInputRole = {
  buyer: "buyer",
  seller: "seller",
  both: "both",
} as const;

export interface GuestLoginInput {
  name: string;
  role: GuestLoginInputRole;
}

export type RegisterInputRole =
  (typeof RegisterInputRole)[keyof typeof RegisterInputRole];

export const RegisterInputRole = {
  buyer: "buyer",
  seller: "seller",
  both: "both",
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
  state?: string;
  city?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type GetListingsParams = {
  category?: string;
  ageGroup?: string;
  state?: string;
  city?: string;
  giftItForward?: boolean;
  search?: string;
  page?: number;
  limit?: number;
};