# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BebeMart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get listings with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: ageGroup
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: giftItForward
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{listingId}:
    get:
      operationId: getListingById
      tags: [listings]
      summary: Get a single listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingInput"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /listings/user/{userId}:
    get:
      operationId: getUserListings
      tags: [listings]
      summary: Get listings for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
...
  /messages/{otherUserId}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get messages with another user
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /auth/guest:
    post:
      operationId: loginAsGuest
      tags: [auth]
      summary: Login or register as guest user (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestLoginInput"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, seller, both]
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - reviewCount
        - createdAt
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - reviewCount
        - createdAt
    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
        ageGroup:
          type: string
        condition:
          type: string
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerRating:
          type: number
          nullable: true
        state:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [available, sold, reserved]
        giftItForward:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - price
        - category
        - ageGroup
        - condition
        - images
        - sellerId
        - sellerName
        - state
        - city
        - status
        - giftItForward
        - createdAt
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/Listing"
        - type: object
          properties:
            seller:
              $ref: "#/components/schemas/UserPublic"
          required:
            - seller
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - totalPages
    CreateListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: integer
          minimum: 0
          maximum: 5000
        category:
          type: string
          enum: [Clothes, Shoes, Toys, Gear, Books]
        ageGroup:
          type: string
          enum: ["0-6m", "6-12m", "1-3y", "4-6y", "7y+"]
        condition:
          type: string
          enum: [Like New, Good, Fair]
        images:
          type: array
          items:
            type: string
          minItems: 1
        videoUrl:
          type: string
          nullable: true
        state:
          type: string
        city:
          type: string
        giftItForward:
          type: boolean
      required:
        - title
        - description
        - price
        - category
        - ageGroup
        - condition
        - images
        - state
        - city
        - giftItForward
    UpdateListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: integer
          minimum: 0
          maximum: 5000
        category:
          type: string
        ageGroup:
          type: string
        condition:
          type: string
        images:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        state:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [available, sold, reserved]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        content:
          type: string
        listingId:
          type: integer
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - senderId
        - receiverId
        - content
        - read
        - createdAt
    SendMessageInput:
      type: object
      properties:
        content:
          type: string
        listingId:
          type: integer
          nullable: true
      required:
        - content
...
    GuestLoginInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [buyer, seller, both]
      required:
        - name
        - role
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [buyer, seller, both]
        state:
          type: string
        city:
          type: string
      required:
        - name
        - email
        - password
        - role
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password