# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BebeMart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputRole } from "./registerInputRole";

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
  state?: string;
  city?: string;
}